CRUD categorias

+++ app/Http/Controllers/Admin/FamilyController.php
@@ -14,7 +14,8 @@ class FamilyController extends Controller
     public function index()
     {
         $families = Family::orderBy('id', 'desc')
-            ->paginate(10);//para visualizar de 10 en 10
+            ->paginate(10)//para visualizar de 10 en 10
+            ->onEachSide(1);
         return view('admin.families.index', compact('families'));
     }
 
@@ -36,6 +37,12 @@ class FamilyController extends Controller
         ]);
 
         Family::create($request->all());
+
+        session()->flash('swal', [
+            'icon' => 'success',
+            'title' => '¡Bien hecho!',
+            'text' => 'Familia creada correctamente'
+        ]);
 
         return redirect()->route('admin.families.index');
     }
@@ -66,6 +73,12 @@ class FamilyController extends Controller
         ]);
 
         $family->update($request->all());
+
+        session()->flash('swal', [
+            'icon' => 'success',
+            'title' => '¡Bien hecho!',
+            'text' => 'Familia actualizada correctamente'
+        ]);
         
         return redirect()->route('admin.families.edit', $family);
     }
@@ -75,7 +88,25 @@ class FamilyController extends Controller
      */
     public function destroy(Family $family)
     {
+
+        if ($family->categories->count() > 0){
+
+            session()->flash('swal', [
+                'icon' => 'error',
+                'title' => '¡Ups!',
+                'text' => 'No se puede eliminar la familia porque tiene categorias asociadas.'
+            ]);
+
+            return redirect()->route('admin.families.edit', $family);
+
+        }
         $family->delete();
+
+        session()->flash('swal', [
+            'icon' => 'success',
+            'title' => 'Bien hecho!',
+            'text' => 'Familia eliminada correctamente.'
+        ]);
 
         return redirect()->route('admin.families.index');
     }
